requirements.txt
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
idna==3.10
imaplib2==3.6
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
